List 3 most recent posts on home page
